refactor(ui): extract new footer and use it on every screen
