+++ packages/nextjs/ui/index.ts
@@ -1,4 +1,5 @@
 export { Button } from "./Button/Button";
+export { Footer } from "./Footer";
 export { Input } from "./Input/Input";
 export { LargeLogo } from "./LargeLogo";
 export { Text } from "./Text/Text";
